auth-service/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2 
PyJWT==2.7.0 
psycopg2-binary==2.9.6
Flask-Cors==4.0.0 